Move wpcom.undocumented().getAtomicSiteMediaViaProxy to a separate module
