+++ client/my-sites/media-library/proxied-image.tsx
@@ -1,7 +1,7 @@
 import debugFactory from 'debug';
 import { useEffect, useState } from 'react';
-import * as React from 'react';
-import wpcom from 'calypso/lib/wp';
+import { getAtomicSiteMediaViaProxyRetry } from 'calypso/lib/get-atomic-site-media';
+import type { ComponentType, FC, ReactElement, ReactNode } from 'react';
 
 const debug = debugFactory( 'calypso:my-sites:media-library:proxied-image' );
 
@@ -9,13 +9,13 @@ type RenderedComponentProps = {
 	src: string;
 	[ key: string ]: any;
 };
-export type RenderedComponent = string | React.ComponentType< RenderedComponentProps >;
+export type RenderedComponent = string | ComponentType< RenderedComponentProps >;
 
 export interface ProxiedImageProps {
 	query: string;
 	filePath: string;
 	siteSlug: string;
-	placeholder: React.ReactNode | null;
+	placeholder: ReactNode;
 	component: RenderedComponent;
 	maxSize: number | null;
 	onError?: ( err: Error ) => any;
@@ -39,7 +39,7 @@ const cacheResponse = ( requestId: string, blob: Blob, freshness = 60000 ) => {
 	}, freshness );
 };
 
-const ProxiedImage: React.FC< ProxiedImageProps > = function ProxiedImage( {
+const ProxiedImage: FC< ProxiedImageProps > = function ProxiedImage( {
 	siteSlug,
 	filePath,
 	query,
@@ -64,9 +64,7 @@ const ProxiedImage: React.FC< ProxiedImageProps > = function ProxiedImage( {
 			if ( maxSize !== null ) {
 				options.maxSize = maxSize;
 			}
-			wpcom
-				.undocumented()
-				.getAtomicSiteMediaViaProxyRetry( siteSlug, filePath, options )
+			getAtomicSiteMediaViaProxyRetry( siteSlug, filePath, options )
 				.then( ( data: Blob ) => {
 					cacheResponse( requestId, data );
 					setImageObjectUrl( URL.createObjectURL( data ) );
@@ -84,7 +82,7 @@ const ProxiedImage: React.FC< ProxiedImageProps > = function ProxiedImage( {
 	}, [ siteSlug, filePath, query ] );
 
 	if ( ! imageObjectUrl ) {
-		return placeholder as React.ReactElement;
+		return placeholder as ReactElement;
 	}
 
 	/* eslint-disable-next-line jsx-a11y/alt-text */
